Move back to services. According to https://developer.android.com/develop/background-work/background-tasks#user-initiated, services are the right API after all, not WorkManager workers. (task initiated by the user (albeit from another app), not ok if deferred, critical task, no other alternative API (TBC) => Foreground service
The important point is to do all necessary stuff to foreground the service while we are still in the binder call that initiates the file save on the SAF client's behalf. If we do it moments too late, SAF client may have closed its activity, depriving us of "BTOP" status needed to start foreground service.

With Workers, it was not possible to do this foregrounding at that right moment.